+++ WIP.js
@@ -1,0 +1,8 @@
+//    windSpeed.stage.x; //how much canvas has been shifted
+//    var context = windSpeed.canvas.getContext('2d');
+//    do {
+//        windSpeed.setupVars.textDisplaySize++;
+//        context.font = "bold " + windSpeed.setupVars.textDisplaySize * 1.5 + "px arial";
+//    } while (context.measureText(windSpeed.textTitle.text).width / 2 < (windSpeed.canvas.width / 2));
+//    
+//    windSpeed.textDisplaySize--;